Fix Cli configuration
